include multi-tenant initialize e2e test
